requirements.txt
# ============================================
# DEPENDENCIAS MÍNIMAS PARA RENDER.COM
# ============================================
# Versión ultra simplificada para compatibilidad con Python 3.13
# Solo dependencias esenciales para el backend básico

# FastAPI y servidor
fastapi==0.95.2
uvicorn[standard]==0.22.0

# Utilidades básicas
python-dotenv==1.0.0
python-multipart==0.0.6

# Cliente HTTP
httpx==0.24.1

# Base de datos - Versión compatible con Python 3.13
sqlalchemy==2.0.23
psycopg2-binary==2.9.6

# Autenticación
PyJWT==2.7.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Utilidades del sistema
psutil==5.9.5
nest_asyncio==1.5.6

# Supabase
supabase==2.3.4 